requirements.txt
tda-api
yfinance
matplotlib
pandas
numpy